+++ include/nginx.sh
@@ -51,6 +51,22 @@ EO_NG_NSL
 
 }
 
+nginx_listen()
+{
+	local _port="${1:-80}"
+	local _opts="${2:-}"
+	if [ -n "$_opts" ]; then _opts=" $_opts"; fi
+
+	get_public_ip4
+	get_public_ip6
+
+	local _c4=""; if [ -z "${PUBLIC_IP4:-}" ]; then _c4="#"; fi
+	local _c6=""; if [ -z "${PUBLIC_IP6:-}" ]; then _c6="#"; fi
+
+	printf '\t\t%slisten       %s%s;\n\t\t%slisten  [::]:%s%s;\n' \
+		"$_c4" "$_port" "$_opts" "$_c6" "$_port" "$_opts"
+}
+
 contains() {
 	string="$1"
 	substring="$2"
@@ -71,13 +87,8 @@ configure_nginx_server_d()
 
 	# no more proxy protocol on backends, since nginx can't
 	# send proxy protocol AND route URIs at the same time
-	local _prefix='server {
-		listen       80;'
-
-	if [ -n "${PUBLIC_IP6:-}" ]; then
-		_prefix="$_prefix
-		listen  [::]:80;"
-	fi
+	local _prefix; _prefix="server {
+$(nginx_listen 80)"
 
 	local _suffix='location ~ /\.ht {
 			deny  all;